refactor(executor): Split morsel.rs into smaller modules
Split the 1894-line morsel.rs file into focused modules:
- mod.rs (142 lines): Core Morsel struct and work-stealing utilities
- config.rs (323 lines): MorselConfig and adaptive sizing
- parallel.rs (502 lines): Filter, map, reduce, group operations
- sort.rs (248 lines): K-way merge sort implementation
- join.rs (147 lines): Hash join probe operations
- tests.rs (603 lines): All unit tests

No public API changes. All 35 existing tests pass.

Closes #4323
